requirements.txt
Django==4.2.3
djangorestframework==3.14.0
drf-spectacular==0.26.3
gunicorn==20.1.0